+++ src/main/java/com/edufy/edufy/repositories/TrackRepository.java
@@ -3,6 +3,6 @@ package com.edufy.edufy.repositories;
 import com.edufy.edufy.models.Track;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrackRepository extends JpaRepository<Track,Long> {
+public interface TrackRepository extends JpaRepository<Track,Integer> {
 
 }
